requirements.txt
hopsworks
xgboost
scikit-learn
pandas
numpy
yfinance
fredapi
newsapi-python
transformers
torch
gradio
python-dotenv
matplotlib
seaborn
plotly
jupyter
ipykernel
pyyaml
tqdm
joblib
